allocator: Old literal tracking
If a register is known to contain a literal, and we need the same literal, we
 can just load it from that register, which is cheaper than a literal load:
LD r, r' is 4T, 1B vs. 7T, 2B for LD r,n
Of course, it's already in the right register, it costs 0T, 0B!
However, we currently have a slow 16-bit move:
PUSH qq; POP qq is 21T, 2B vs 10T, 3B for LD dd,nn
If we replaced this with a pair of 8-bit moves it would be 8T, 2B.
So let's do that.